minor restructuring in crab_template
now by default the psetName which holds the path of the ntuplewriter and the outLFNDirBase is chosen per dataset. The User still can overwrite this by additionally enabling two respective flags.
